pyproject.toml
[project]
name = "hikayeapp-backend"
version = "0.1.0"
description = "Story teller AI application backend services"
authors = [
    {name = "recep.daban@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "python-multipart (>=0.0.20, <0.0.21)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "google-auth (>=2.40.3,<3.0.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "cryptography (>=45.0.6,<46.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "bcrypt (==4.0.1)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-openai (>=0.3.30,<0.4.0)",
]

[tool.poetry]
packages = [{ include = "hikayeapp_backend" }]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
